split-utils * Move useFormFields to hooks/useFormFields.ts
